requirements.txt
Flask==3.0.3
openai==1.54.4
python-dotenv==1.0.1
Werkzeug==3.1.3
gunicorn==23.0.0
